fix: prevent major bug that could potentially remove $HOME/.gnupg
In the rare (but eventual) case where the user has $GNUPGHOME environment variable set and
$USE_SYSTEM_GPG turned on (for legacy compatibility reasons), pacom would remove their $GNUPGHOME
without asking questions.

Notice that by default, as of v2, $USE_SYSTEM_GPG is OFF, so this bug wouldn't hit us right off the
bat.

Also, this fixes a second bug where your entire GNUPG public vault could be leaked to your git-repo
when you use `pacom upgrade`

+++ lib/gpg.sh
@@ -2,14 +2,15 @@
 # Common GPG-related functions
 
 function gpg::start {
-	# Do nothing if we're using system GPG
-	test -n "$USE_SYSTEM_GPG" && return
+	# Do nothing if we're using system GPG or already initialized
+	[ -n "$USE_SYSTEM_GPG" ] || [ "$_gpg_initialized" = 1 ] && return
 
 	# Create the GPG path
 	msg "Mounting a temporary GPG home structure"
 	export GNUPGHOME="$GIT_REPO_PATH/.gnupg"
-	test -d "$GNUPGHOME" || mkdir "$GNUPGHOME"
-	chmod 700 "$GNUPGHOME"
+	# Cleanup if needed
+	test -d "$GNUPGHOME" && rm -rf "$GNUPGHOME"
+	mkdir "$GNUPGHOME" && chmod 700 "$GNUPGHOME"
 
 	# Then import saved keys onto it
 	local keys_path; keys_path="$(gpg::keys_path)"
@@ -19,15 +20,19 @@ function gpg::start {
 
 	# Ensure we don't get pubring.kbx creation stderr messages
 	gpg -k > /dev/null 2>&1
-}
 
-function gpg::keys_path {
-	echo "$GIT_REPO_PATH/gpg-keys.asc"
+	# Set a flag that says the GPG infrastructure has been initialized
+	# This prevents calls to gpg::sync_db to backup the system keyring instead of Pacom's when it
+	# hasn't been initialized yet
+	_gpg_initialized=1
+
+	# Make sure we cleanup on exit
+	trap gpg::_cleanup EXIT
 }
 
 function gpg::sync_db {
-	# Do nothing if we're using system gpg
-	test -n "$USE_SYSTEM_GPG" && return
+	# Do nothing if we're using system gpg (or haven't initialized using gpg::start)
+	[ -n "$USE_SYSTEM_GPG" ] || [ "$_gpg_initialized" != 1 ] && return
 
 	# Save all existing keys onto our git repo
 	local keys_path; keys_path="$(gpg::keys_path)"
@@ -39,7 +44,16 @@ function gpg::sync_db {
 	git::commit_gpg ":arrow_up: update gpg database"
 }
 
-function gpg::cleanup {
-	msg "Cleaning up temporary GPG home structure"
+# Retrieve the path to the gpg keys file on the git repo
+function gpg::keys_path {
+	echo "$GIT_REPO_PATH/gpg-keys.asc"
+}
+
+# Cleanup temporary $GNUPGHOME created by pacom
+# This function should not be called manually
+function gpg::_cleanup {
+	# Do nothing if we're using system gpg or we're not using Pacom's gpg home
+	[ -n "$USE_SYSTEM_GPG" ] || [ "$_gpg_initialized" != 1 ] && return
+
 	rm -rf "$GNUPGHOME"
 }
